When bootstrapping Bazel build in batch mode.
It currently is impossible to build Bazel using ./compile.sh in a chroot
or Docker image.  That's unfortunate cause it would make bug
documentation and reproduction much easier. This patch switches the
bootstrap.sh script to build using "batch" mode. Batch mode "provides
more predictable semantics with respect to signal handling, job control,
and environment variable inheritance, and is necessary for running Bazel
in a chroot jail".

Interestingly, this doesn't actually fix the problem (still working on
that), but it does make it fail instead of looping forever trying to
terminate an old Bazel server. Fortunately it fails after it has
produced a Bazel binary which at least lets me choose to ignore it and
move on.

Bug: https://github.com/bazelbuild/bazel/issues/134
Change-Id: Ia008af973f9e63836826dfe65f10e5fbf4967bc6